Replace better_profanity with regex checking
Since the module fails on some really basic inputs and doesn't support
ignoring word bounds, which we now do using asterisks. We don't try to
recognize users trying to circumvent it, since that's an offense we
punish manually. We also use an edit wordlist from it as the default,
since the one it provides is kinda questionable.